implement MNISTDataset struct for Array, move functions
